+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Make sure QrCode `min_amount` and `max_amount` make sense
+
 ### Removed
 - Removed `phone2` and `addPhoneNumbers()` from `Customer` class